% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptos_2017to2021.R
\docType{data}
\name{cryptos_2017to2021}
\alias{cryptos_2017to2021}
\title{Cryptocurrency daily and hourly price data between 2017 and 2021}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(cryptos_2017to2021)
}
\description{
Cryptocurrency daily (since 2017) and hourly (since 2021) price data between 2017 and 2021.
}
\keyword{datasets}
