% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP500_2015to2020.R
\docType{data}
\name{SP500_2015to2020}
\alias{SP500_2015to2020}
\title{Stock price data of the index and constituent stocks of the S&P 500 from 2015 to 2020}
\format{
An object of class \code{list} of length 2.
}
\source{
Yahoo!Finance, Wikipedia (https://en.wikipedia.org/wiki/List_of_S%26P_500_companies)
}
\usage{
data(SP500_2015to2020)
}
\description{
Stock data downloaded from Yahoo!Finance and cleaned with package `imputeFin` (for missing values and outliers).
All the data is adjusted for splits & dividends.
}
\keyword{datasets}
