% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptos_2017to2021.R
\docType{data}
\name{cryptos_2017to2021}
\alias{cryptos_2017to2021}
\title{Cryptocurrency daily and hourly price data between 2017 and 2021.}
\format{
List with the following elements:
\describe{
  \item{daily}{An `xts` object from 2017 to 2021 with daily close prices of 9 cryptocurrencies.}
  \item{hourly}{An `xts` object from 2021-01 to 2021-06 with hourly close prices of 82 cryptocurrencies.}
}
}
\usage{
data(cryptos_2017to2021)
}
\description{
Cryptocurrency daily data (since 2017) and hourly data (since 2021) until 2021.
}
\keyword{datasets}
